plugins/simple_django_dashboard/0.0.1/pyproject.toml
[build-system]
requires = ['setuptools', 'setuptools-scm']
build-backend = 'setuptools.build_meta'

[project]
name = 'simple_django_dashboard'
version = '0.0.1'
authors = [
    {name = 'Vladimir Molodtsov', email = 'vmol@mail.ru'},
]
dependencies = [
    'django',
    'pyzmq',
]